requirements.txt
Flask==2.1.3
pandas==1.5.2
pdfminer.six
unidecode==1.3.6
openpyxl
requests==2.28.1
Werkzeug==2.2.3
itsdangerous==2.1.2
